requirements.txt
Django>=4.2
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-js-asset==3.0.1
django_csp==3.8
pillow==11.1.0
python-dotenv==1.0.1
whitenoise==6.8.2
django-check-seo==1.0.1
django-unfold==0.46.0
django-tailwind==3.8.0
django-axes==7.0.1
django-dbbackup==4.2.1
bleach==6.2.0
redis==5.2.1
sqlparse==0.5.3
tzdata==2024.2
webencodings==0.5.1
gunicorn==23.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1 